Added UpdateTask to replace the deprecated LaunchActivity.BGLogin class Changed all update references to use the Broadcast class Fixed some timestamps using int instead of long in the Gson/SnapChat compat classes

+++ OpenSnap/src/com/nickstephen/opensnap/util/http/ServerFriend.java
@@ -10,7 +10,7 @@ public class ServerFriend {
     public String name;
     public String display;
     public int type;
-    public int ts;
+    public long ts;
     public List<String> best_friends;
     public int score;
